Rearranged DB code, seperate files for models
